Add docs and modify no_update hook
The previous version of the no_update hook was implemented such
that the hooks ran no matter what error was thrown, despite the
plan being to only run when there was no update to be made, but
the API call was otherwise successful.

This commit also documents the feature and adds some minor changes
to other parts of the hooks.md page.